flask blueprint + dupe asset cleanup
