+++ app/api/__init__.py
@@ -1,0 +1,2 @@
+from .resources import api
+
