+++ run.py
@@ -1,0 +1,4 @@
+from flask import Flask
+from app import create_app
+
+http_server = create_app(Flask)
