--- static/cropped_tweet_files/twitter_more_2.css
@@ -1,425 +1,0 @@
-@keyframes underOverNavBar {
-  0% {
-    top: -50px;
-    z-index: 999;
-  }
-
-  50% {
-    top: 0;
-    z-index: 999;
-  }
-
-  70% {
-    top: 47px;
-    z-index: 999;
-  }
-
-  80% {
-    top: 65px;
-    z-index: 999;
-  }
-
-  84% {
-    top: 65px;
-    z-index: 1001;
-  }
-
-  100% {
-    top: 47px;
-    z-index: 1001;
-  }
-}
-
-@keyframes ContactImportButtonsTransitionOut {
-  0% {
-    opacity: 1;
-    -ms-filter: "alpha(opacity=100)";
-    visibility: visible;
-    position: absolute;
-  }
-
-  50% {
-    opacity: 0;
-    -ms-filter: "alpha(opacity=0)";
-    visibility: hidden;
-    position: absolute;
-  }
-}
-
-@keyframes ContactImportButtonsTransitionIn {
-  0% {
-    visibility: hidden;
-    position: absolute;
-    opacity: 0;
-    -ms-filter: "alpha(opacity=0)";
-  }
-
-  50.5% {
-    opacity: 0;
-    -ms-filter: "alpha(opacity=0)";
-    visibility: visible;
-    position: static;
-  }
-
-  100% {
-    opacity: 1;
-    -ms-filter: "alpha(opacity=100)";
-  }
-}
-
-@keyframes StartCongratulationsFadeIn {
-  0% {
-    opacity: 0;
-    -ms-filter: "alpha(opacity=0)";
-    transform: translateY(10px);
-  }
-
-  100% {
-    opacity: 1;
-    -ms-filter: "alpha(opacity=100)";
-    transform: translateY(0px);
-  }
-}
-
-.Avatar {
-  -moz-force-broken-image-icon: 1;
-  display: block;
-  border-radius: 50%;
-}
-
-.Avatar--size48 {
-  height: 48px;
-  width: 48px;
-}
-
-.Avatar--size32 {
-  height: 32px;
-  width: 32px;
-}
-
-@keyframes ProfileHeaderCard-periscopeProfileBlink {
-  0% {
-    color: #657786;
-  }
-
-  40% {
-    color: #e0245e;
-  }
-
-  60% {
-    color: #e0245e;
-  }
-
-  100% {
-    color: #657786;
-  }
-}
-
-.dropdown-form {
-  padding: 15px;
-  left: auto;
-  right: 0;
-}
-
-.dropdown-form label {
-  color: #14171a;
-}
-
-.dropdown-form p a {
-  display: inline;
-}
-
-.mobile-complete {
-  padding: 10px 0;
-  border-radius: 5px;
-}
-
-.dropdown-menu {
-  width: auto;
-}
-
-.dropdown-menu .button {
-  float: right;
-}
-
-.dropdown-menu input[type=text],.dropdown-menu input[type=password] {
-  width: 230px;
-}
-
-.link-form {
-  display: inline;
-}
-
-.link-form .btn-link {
-  font-size: 14px;
-  color: #1c94e0;
-}
-
-@keyframes MomentAnimation-keyframes {
-  0% {
-    background-position: left;
-  }
-
-  100% {
-    background-position: right;
-  }
-}
-
-/*! jQuery UI - v1.8.22 - 2012-07-24
-* https://github.com/jquery/jquery-ui
-* Includes: jquery.ui.core.css, jquery.ui.accordion.css, jquery.ui.autocomplete.css, jquery.ui.button.css, jquery.ui.datepicker.css, jquery.ui.dialog.css, jquery.ui.progressbar.css, jquery.ui.resizable.css, jquery.ui.selectable.css, jquery.ui.slider.css, jquery.ui.tabs.css, jquery.ui.theme.css
-* Copyright (c) 2012 AUTHORS.txt; Licensed MIT, GPL */
-
-.ui-icon {
-  display: block;
-  text-indent: -99999px;
-  overflow: hidden;
-  background-repeat: no-repeat;
-}
-
-.ui-menu {
-  list-style: none;
-  padding: 2px;
-  margin: 0;
-  display: block;
-  float: left;
-}
-
-.ui-menu .ui-menu {
-  margin-top: -3px;
-}
-
-.ui-menu .ui-menu-item {
-  margin: 0;
-  padding: 0;
-  zoom: 1;
-  float: left;
-  clear: left;
-  width: 100%;
-}
-
-.ui-menu .ui-menu-item a {
-  text-decoration: none;
-  display: block;
-  padding: .2em .4em;
-  line-height: 1.5;
-  zoom: 1;
-}
-
-.ui-menu .ui-menu-item a.ui-state-hover {
-  font-weight: normal;
-  margin: -1px;
-}
-
-.ui-button {
-  display: inline-block;
-  position: relative;
-  padding: 0;
-  margin-right: .1em;
-  text-decoration: none!important;
-  cursor: pointer;
-  text-align: center;
-  zoom: 1;
-  overflow: visible;
-}
-
-.ui-button-icon-only {
-  width: 2.2em;
-}
-
-button.ui-button-icon-only {
-  width: 2.4em;
-}
-
-.ui-button-icons-only {
-  width: 3.4em;
-}
-
-button.ui-button-icons-only {
-  width: 3.7em;
-}
-
-.ui-button .ui-button-text {
-  display: block;
-  line-height: 1.4;
-}
-
-.ui-button-text-only .ui-button-text {
-  padding: .4em 1em;
-}
-
-.ui-button-icon-only .ui-button-text,.ui-button-icons-only .ui-button-text {
-  padding: .4em;
-  text-indent: -9999999px;
-}
-
-.ui-button-text-icon-primary .ui-button-text,.ui-button-text-icons .ui-button-text {
-  padding: .4em 1em .4em 2.1em;
-}
-
-.ui-button-text-icon-secondary .ui-button-text,.ui-button-text-icons .ui-button-text {
-  padding: .4em 2.1em .4em 1em;
-}
-
-.ui-button-text-icons .ui-button-text {
-  padding-left: 2.1em;
-  padding-right: 2.1em;
-}
-
-input.ui-button {
-  padding: .4em 1em;
-}
-
-.ui-button-icon-only .ui-icon,.ui-button-text-icon-primary .ui-icon,.ui-button-text-icon-secondary .ui-icon,.ui-button-text-icons .ui-icon,.ui-button-icons-only .ui-icon {
-  position: absolute;
-  top: 50%;
-  margin-top: -8px;
-}
-
-.ui-button-icon-only .ui-icon {
-  left: 50%;
-  margin-left: -8px;
-}
-
-.ui-button-text-icon-primary .ui-button-icon-primary,.ui-button-text-icons .ui-button-icon-primary,.ui-button-icons-only .ui-button-icon-primary {
-  left: .5em;
-}
-
-.ui-button-text-icon-secondary .ui-button-icon-secondary,.ui-button-text-icons .ui-button-icon-secondary,.ui-button-icons-only .ui-button-icon-secondary {
-  right: .5em;
-}
-
-.ui-button-text-icons .ui-button-icon-secondary,.ui-button-icons-only .ui-button-icon-secondary {
-  right: .5em;
-}
-
-button.ui-button::-moz-focus-inner {
-  border: 0;
-  padding: 0;
-}
-
-.ui-state-default {
-  border: 1px solid #d3d3d3;
-  background: #e6e6e6 url(images/ui-bg_glass_75_e6e6e6_1x400.png) 50% 50% repeat-x;
-  font-weight: normal;
-  color: #66757f;
-}
-
-.ui-state-default a,.ui-state-default a:link,.ui-state-default a:visited {
-  color: #66757f;
-  text-decoration: none;
-}
-
-.ui-state-hover,.ui-state-focus {
-  border: 1px solid #657786;
-  background: #dadada url(images/ui-bg_glass_75_dadada_1x400.png) 50% 50% repeat-x;
-  font-weight: normal;
-  color: #212121;
-}
-
-.ui-state-hover a,.ui-state-hover a:hover {
-  color: #212121;
-  text-decoration: none;
-}
-
-.ui-state-error {
-  border: 1px solid #cd0a0a;
-  background: #fef1ec url(images/ui-bg_glass_95_fef1ec_1x400.png) 50% 50% repeat-x;
-  color: #cd0a0a;
-}
-
-.ui-state-error a {
-  color: #cd0a0a;
-}
-
-.ui-state-error-text {
-  color: #cd0a0a;
-}
-
-.ui-icon {
-  width: 16px;
-  height: 16px;
-  background-image: url(images/ui-icons_222222_256x240.png);
-}
-
-.ui-state-default .ui-icon {
-  background-image: url(images/ui-icons_888888_256x240.png);
-}
-
-.ui-state-hover .ui-icon,.ui-state-focus .ui-icon {
-  background-image: url(images/ui-icons_454545_256x240.png);
-}
-
-.ui-state-error .ui-icon,.ui-state-error-text .ui-icon {
-  background-image: url(images/ui-icons_cd0a0a_256x240.png);
-}
-
-.ui-icon-document {
-  background-position: -32px -96px;
-}
-
-.ui-icon-document-b {
-  background-position: -48px -96px;
-}
-
-.ui-icon-search {
-  background-position: -160px -112px;
-}
-
-.ui-icon-heart {
-  background-position: -208px -112px;
-}
-
-.ui-icon-link {
-  background-position: -240px -112px;
-}
-
-.ui-icon-cancel {
-  background-position: 0 -128px;
-}
-
-.ui-icon-key {
-  background-position: -112px -128px;
-}
-
-.ui-icon-copy {
-  background-position: -176px -128px;
-}
-
-.ui-icon-image {
-  background-position: -208px -128px;
-}
-
-.ui-icon-script {
-  background-position: -240px -128px;
-}
-
-/*!
- * jQuery UI Slider 1.8.22
- *
- * Copyright 2012, AUTHORS.txt (http://jqueryui.com/about)
- * Dual licensed under the MIT or GPL Version 2 licenses.
- * http://jquery.org/license
- *
- * http://docs.jquery.com/UI/Slider#theming
- */
-
-.profile-avatar-modal {
-  width: 350px;
-}
-
-.profile-header-modal {
-  width: 630px;
-}
-
-@keyframes expandBirthdateFields {
-  0% {
-    max-height: 35px;
-  }
-
-  100% {
-    max-height: 300px;
-    opacity: 1;
-    -ms-filter: "alpha(opacity=100)";
-  }
-}
