+++ templates/index.html
@@ -40,7 +40,7 @@
                 </form>
             </div>
             <div class="col">
-                <img src="/static/key.png" class="img-fluid" alt="Key">
+		<img src="{{ url_for('static', filename='key.png') }}" class="img-fluid" alt="Key">
             </div>
 
         </div>
